day_2/task_1/requirements.txt
chromadb 
PyMuPDF 
google-generativeai 
python-dotenv 
tqdm 
sentence-transformers